Move tests from separate package
