+++ controllers/api_controller_integration_test.go
@@ -703,10 +703,9 @@ var _ = Describe("APIRule Controller", Serial, func() {
 					}()
 
 					expectApiRuleStatus(apiRuleName, gatewayv1beta1.StatusOK)
-					updateJwtHandlerTo(helpers.JWT_HANDLER_ISTIO)
 
 					// when
-					triggerApiRuleReconciliation(apiRuleName)
+					updateJwtHandlerTo(helpers.JWT_HANDLER_ISTIO)
 
 					// then
 					Eventually(func(g Gomega) {
@@ -737,7 +736,6 @@ var _ = Describe("APIRule Controller", Serial, func() {
 
 					expectApiRuleStatus(apiRuleName, gatewayv1beta1.StatusOK)
 					updateJwtHandlerTo(helpers.JWT_HANDLER_ISTIO)
-					triggerApiRuleReconciliation(apiRuleName)
 					expectApiRuleStatus(apiRuleName, gatewayv1beta1.StatusError)
 
 					// when
@@ -783,10 +781,9 @@ var _ = Describe("APIRule Controller", Serial, func() {
 
 					By("Waiting until reconciliation of API Rule is finished")
 					expectApiRuleStatus(apiRuleName, gatewayv1beta1.StatusOK)
-					updateJwtHandlerTo(helpers.JWT_HANDLER_ORY)
 
 					// when
-					triggerApiRuleReconciliation(apiRuleName)
+					updateJwtHandlerTo(helpers.JWT_HANDLER_ORY)
 
 					// then
 					Eventually(func(g Gomega) {
@@ -819,7 +816,6 @@ var _ = Describe("APIRule Controller", Serial, func() {
 					By("Waiting until reconciliation of API Rule is finished")
 					expectApiRuleStatus(apiRuleName, gatewayv1beta1.StatusOK)
 					updateJwtHandlerTo(helpers.JWT_HANDLER_ORY)
-					triggerApiRuleReconciliation(apiRuleName)
 
 					By("Waiting until reconciliation of API Rule is finished")
 					Eventually(func(g Gomega) {
@@ -850,6 +846,68 @@ var _ = Describe("APIRule Controller", Serial, func() {
 				})
 			})
 		})
+	})
+
+	It("APIRule in status Error should reconcile to status OK when root cause of error is fixed", func() {
+		// given
+		updateJwtHandlerTo(helpers.JWT_HANDLER_ISTIO)
+
+		apiRuleName := generateTestName(testNameBase, testIDLength)
+		serviceName := generateTestName(testServiceNameBase, testIDLength)
+		serviceHost := fmt.Sprintf("%s.kyma.local", serviceName)
+		vsName := generateTestName("duplicated-host-vs", testIDLength)
+
+		By(fmt.Sprintf("Create Virtual service for host %s", serviceHost))
+		vs := virtualService(vsName, serviceHost)
+		Expect(c.Create(context.TODO(), vs)).Should(Succeed())
+		defer func() {
+			By(fmt.Sprintf("Deleting VirtualService %s as part of teardown", vs.Name))
+			Eventually(func(g Gomega) {
+				_ = c.Delete(context.TODO(), vs)
+				v := networkingv1beta1.VirtualService{}
+				err := c.Get(context.TODO(), client.ObjectKey{Name: vs.Name, Namespace: testNamespace}, &v)
+				g.Expect(errors.IsNotFound(err)).To(BeTrue())
+			}, eventuallyTimeout).Should(Succeed())
+		}()
+
+		By("Verify VirtualService is created")
+		Eventually(func(g Gomega) {
+			createdVs := networkingv1beta1.VirtualService{}
+			g.Expect(c.Get(context.TODO(), client.ObjectKey{Name: vsName, Namespace: testNamespace}, &createdVs)).Should(Succeed())
+		}, eventuallyTimeout).Should(Succeed())
+
+		apiRuleLabelMatcher := matchingLabelsFunc(apiRuleName, testNamespace)
+
+		By("Create APIRule")
+		rule := testRule("/headers", []string{"GET"}, nil, testIstioJWTHandler(testIssuer, testJwksUri))
+		instance := testInstance(apiRuleName, testNamespace, serviceName, serviceHost, testServicePort, []gatewayv1beta1.Rule{rule})
+
+		Expect(c.Create(context.TODO(), instance)).Should(Succeed())
+		defer func() {
+			deleteApiRule(instance)
+		}()
+
+		expectApiRuleStatus(apiRuleName, gatewayv1beta1.StatusError)
+
+		By("Verify VirtualService for APIRule is not created")
+		Eventually(func(g Gomega) {
+			vsList := networkingv1beta1.VirtualServiceList{}
+			g.Expect(c.List(context.TODO(), &vsList, apiRuleLabelMatcher)).Should(Succeed())
+			g.Expect(vsList.Items).To(HaveLen(0))
+		}, eventuallyTimeout).Should(Succeed())
+
+		By("Deleting existing VirtualService with duplicated host configuration")
+		deleteVirtualService(vs)
+
+		By("Waiting until APIRule is reconciled after error")
+		expectApiRuleStatus(apiRuleName, gatewayv1beta1.StatusOK)
+
+		By("Verify VirtualService for APIRule is created")
+		Eventually(func(g Gomega) {
+			vsList := networkingv1beta1.VirtualServiceList{}
+			g.Expect(c.List(context.TODO(), &vsList, apiRuleLabelMatcher)).Should(Succeed())
+			g.Expect(vsList.Items).To(HaveLen(1))
+		}, eventuallyTimeout).Should(Succeed())
 	})
 })
 
@@ -1089,19 +1147,6 @@ func matchingLabelsFunc(apiRuleName, namespace string) client.ListOption {
 	return client.MatchingLabels(labels)
 }
 
-func triggerApiRuleReconciliation(apiRuleName string) {
-	By("Trigger Reconcile of ApiRule")
-	reconciledApiRule := gatewayv1beta1.APIRule{}
-	Expect(c.Get(context.TODO(), client.ObjectKey{Name: apiRuleName, Namespace: testNamespace}, &reconciledApiRule)).Should(Succeed())
-	// We check for a different generation before reconciliation of APiRule, therefore we need to do a dummy change to
-	// trigger the reconciliation of the ApiRule.
-	newDummyHost := fmt.Sprintf("dummy-change.%s", *reconciledApiRule.Spec.Host)
-	reconciledApiRule.Spec.Host = &newDummyHost
-
-	Expect(c.Update(context.TODO(), &reconciledApiRule)).Should(Succeed())
-
-}
-
 func updateJwtHandlerTo(jwtHandler string) {
 	cm := &corev1.ConfigMap{}
 	Expect(c.Get(context.TODO(), client.ObjectKey{Name: helpers.CM_NAME, Namespace: helpers.CM_NS}, cm)).Should(Succeed())
@@ -1169,4 +1214,25 @@ func expectApiRuleStatus(apiRuleName string, statusCode gatewayv1beta1.StatusCod
 		g.Expect(expectedApiRule.Status.APIRuleStatus.Code).To(Equal(statusCode))
 	}, eventuallyTimeout).Should(Succeed())
 	By(fmt.Sprintf("Validated that ApiRule %s has status %s", apiRuleName, statusCode))
+}
+
+func deleteVirtualService(vs *networkingv1beta1.VirtualService) {
+	By(fmt.Sprintf("Deleting VirtualService %s", vs.Name))
+	Expect(c.Delete(context.TODO(), vs)).Should(Succeed())
+	Eventually(func(g Gomega) {
+		v := networkingv1beta1.VirtualService{}
+		err := c.Get(context.TODO(), client.ObjectKey{Name: vs.Name, Namespace: testNamespace}, &v)
+		g.Expect(errors.IsNotFound(err)).To(BeTrue())
+	}, eventuallyTimeout).Should(Succeed())
+}
+
+func virtualService(name string, host string) *networkingv1beta1.VirtualService {
+	vs := &networkingv1beta1.VirtualService{}
+	vs.ObjectMeta = metav1.ObjectMeta{
+		Name:      name,
+		Namespace: testNamespace,
+	}
+	vs.Spec.Hosts = []string{host}
+
+	return vs
 }
