--- controllers/reconciliation_error_test/suite_test.go
@@ -1,202 +1,0 @@
-package reconciliation_test
-
-import (
-	"context"
-	"fmt"
-	"os"
-	"path/filepath"
-	"testing"
-	"time"
-
-	rulev1alpha1 "github.com/ory/oathkeeper-maester/api/v1alpha1"
-	"istio.io/api/networking/v1beta1"
-	networkingv1beta1 "istio.io/client-go/pkg/apis/networking/v1beta1"
-	securityv1beta1 "istio.io/client-go/pkg/apis/security/v1beta1"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	ctrl "sigs.k8s.io/controller-runtime"
-
-	gatewayv1beta1 "github.com/kyma-project/api-gateway/api/v1beta1"
-	"github.com/kyma-project/api-gateway/controllers"
-	"github.com/kyma-project/api-gateway/internal/helpers"
-	"github.com/kyma-project/api-gateway/internal/processing"
-
-	. "github.com/onsi/ginkgo/v2"
-	"github.com/onsi/ginkgo/v2/reporters"
-	"github.com/onsi/ginkgo/v2/types"
-	. "github.com/onsi/gomega"
-
-	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	// +kubebuilder:scaffold:imports
-)
-
-var (
-	cfg       *rest.Config
-	k8sClient client.Client
-	testEnv   *envtest.Environment
-	requests  chan reconcile.Request
-	c         client.Client
-	ctx       context.Context
-	cancel    context.CancelFunc
-
-	TestAllowOrigins = []*v1beta1.StringMatch{{MatchType: &v1beta1.StringMatch_Regex{Regex: ".*"}}}
-	TestAllowMethods = []string{"GET", "POST", "PUT", "DELETE"}
-	TestAllowHeaders = []string{"header1", "header2"}
-)
-
-func TestAPIGatewayReconciliation(t *testing.T) {
-	RegisterFailHandler(Fail)
-
-	RunSpecs(t, "Reconciliation Suite")
-}
-
-var _ = BeforeSuite(func(specCtx SpecContext) {
-	logf.SetLogger(zap.New(zap.UseDevMode(true), zap.WriteTo(GinkgoWriter)))
-	ctx, cancel = context.WithCancel(context.TODO())
-
-	By("bootstrapping test environment for reconciliation")
-	testEnv = &envtest.Environment{
-		CRDDirectoryPaths: []string{filepath.Join("..", "..", "config", "crd", "bases"), filepath.Join("..", "..", "hack")},
-	}
-
-	var err error
-	cfg, err = testEnv.Start()
-	Expect(err).ToNot(HaveOccurred())
-	Expect(cfg).ToNot(BeNil())
-
-	k8sClient, err = client.New(cfg, client.Options{Scheme: scheme.Scheme})
-	Expect(err).ToNot(HaveOccurred())
-	Expect(k8sClient).ToNot(BeNil())
-
-	s := runtime.NewScheme()
-
-	err = gatewayv1beta1.AddToScheme(s)
-	Expect(err).NotTo(HaveOccurred())
-
-	err = rulev1alpha1.AddToScheme(s)
-	Expect(err).NotTo(HaveOccurred())
-
-	err = networkingv1beta1.AddToScheme(s)
-	Expect(err).NotTo(HaveOccurred())
-
-	err = securityv1beta1.AddToScheme(s)
-	Expect(err).NotTo(HaveOccurred())
-
-	err = corev1.AddToScheme(s)
-	Expect(err).NotTo(HaveOccurred())
-
-	mgr, err := ctrl.NewManager(cfg, ctrl.Options{
-		Scheme:             s,
-		MetricsBindAddress: "0",
-	})
-	Expect(err).NotTo(HaveOccurred())
-
-	c, err = client.New(cfg, client.Options{Scheme: s})
-	Expect(err).NotTo(HaveOccurred())
-
-	ns := &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{Name: testNamespace},
-		Spec:       corev1.NamespaceSpec{},
-	}
-	err = c.Create(context.TODO(), ns)
-	Expect(err).NotTo(HaveOccurred())
-
-	nsKyma := &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{Name: helpers.CM_NS},
-		Spec:       corev1.NamespaceSpec{},
-	}
-	err = c.Create(context.TODO(), nsKyma)
-	Expect(err).NotTo(HaveOccurred())
-
-	cm := &corev1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      helpers.CM_NAME,
-			Namespace: helpers.CM_NS,
-		},
-		Data: map[string]string{
-			helpers.CM_KEY: fmt.Sprintf("jwtHandler: %s", helpers.JWT_HANDLER_ORY),
-		},
-	}
-	err = c.Create(context.TODO(), cm)
-	Expect(err).NotTo(HaveOccurred())
-
-	apiReconciler := &controllers.APIRuleReconciler{
-		Client:            mgr.GetClient(),
-		Log:               ctrl.Log.WithName("controllers").WithName("Api"),
-		OathkeeperSvc:     testOathkeeperSvcURL,
-		OathkeeperSvcPort: testOathkeeperPort,
-		DomainAllowList:   []string{"bar", "kyma.local"},
-		CorsConfig: &processing.CorsConfig{
-			AllowOrigins: TestAllowOrigins,
-			AllowMethods: TestAllowMethods,
-			AllowHeaders: TestAllowHeaders,
-		},
-		GeneratedObjectsLabels: map[string]string{},
-		Config:                 &helpers.Config{},
-
-		// Run the suite with small period so reconcilation should happen very often
-		ReconcilePeriod:        time.Second,
-		OnErrorReconcilePeriod: time.Second,
-	}
-	Expect(err).NotTo(HaveOccurred())
-
-	err = apiReconciler.SetupWithManager(mgr)
-	Expect(err).NotTo(HaveOccurred())
-
-	go func() {
-		defer GinkgoRecover()
-		err = mgr.Start(ctx)
-		Expect(err).ToNot(HaveOccurred(), "failed to run manager")
-	}()
-
-}, NodeTimeout(60*time.Second))
-
-var _ = AfterSuite(func() {
-	/*
-		 Provided solution for timeout issue waiting for kubeapiserver
-			https://github.com/kubernetes-sigs/controller-runtime/issues/1571#issuecomment-1005575071
-	*/
-	cancel()
-
-	By("tearing down the test environment,but I do nothing here.")
-	err := testEnv.Stop()
-	// Set 4 with random
-	if err != nil {
-		time.Sleep(4 * time.Second)
-	}
-	err = testEnv.Stop()
-	Expect(err).NotTo(HaveOccurred())
-
-})
-
-var _ = ReportAfterSuite("custom reporter", func(report types.Report) {
-	logger := zap.New(zap.UseDevMode(true), zap.WriteTo(GinkgoWriter))
-	if key, ok := os.LookupEnv("ARTIFACTS"); ok {
-		reportsFilename := fmt.Sprintf("%s/%s", key, "junit-reconciliation.xml")
-		logger.Info("Generating reports at", "location", reportsFilename)
-		err := reporters.GenerateJUnitReport(report, reportsFilename)
-
-		if err != nil {
-			logger.Error(err, "Junit Report Generation Error")
-		}
-	} else {
-		if err := os.MkdirAll("../../reports", 0755); err != nil {
-			logger.Error(err, "could not create directory")
-		}
-
-		reportsFilename := fmt.Sprintf("%s/%s", "../../reports", "junit-reconciliation.xml")
-		logger.Info("Generating reports at", "location", reportsFilename)
-		err := reporters.GenerateJUnitReport(report, reportsFilename)
-
-		if err != nil {
-			logger.Error(err, "Junit Report Generation Error")
-		}
-	}
-})
